test(ts-browser): move test util into `window.helpers`
This one looks scary but the only thing it does is moving the natively
defined helper functions into the `TestHelpers` class which is defined
in browser.

All helper methods must now be called from within the browser session
via `window.testHelpers`. This avoids data passing between native and
browser context (note that all `browser.execute` calls are gone). This
also allows us to pass cc instances directly as parameters instead of
passing strings and retrieving them from the window.

Overall, this should bring the util and test logic closer to what we
would do in native code.
